+++ FirebaseAuth/Sources/Swift/Utilities/AuthWebViewController.swift
@@ -52,7 +52,20 @@
 
     private var url: URL
     weak var delegate: AuthWebViewControllerDelegate?
-    private weak var webView: AuthWebView?
+
+    private let webView: WKWebView = {
+      let webView = WKWebView(frame: .zero)
+      webView.isOpaque = false
+      webView.backgroundColor = .clear
+      webView.scrollView.isOpaque = false
+      webView.scrollView.backgroundColor = .clear
+      webView.scrollView.bounces = false
+      webView.scrollView.alwaysBounceVertical = false
+      webView.scrollView.alwaysBounceHorizontal = false
+      return webView
+    }()
+
+    private let spinner = UIActivityIndicatorView(style: .medium)
 
     // MARK: - Initialization
 
@@ -69,11 +82,14 @@
 
     // MARK: - View Lifecycle
 
-    override func loadView() {
-      let webView = AuthWebView(frame: UIScreen.main.bounds)
-      webView.webView.navigationDelegate = self
-      view = webView
-      self.webView = webView
+    override func viewDidLoad() {
+      super.viewDidLoad()
+      view.backgroundColor = .white
+
+      webView.navigationDelegate = self
+      view.addSubview(webView)
+      view.addSubview(spinner)
+
       navigationItem.leftBarButtonItem = UIBarButtonItem(
         barButtonSystemItem: .cancel,
         target: self,
@@ -81,9 +97,15 @@
       )
     }
 
+    override func viewDidLayoutSubviews() {
+      super.viewDidLayoutSubviews()
+      webView.frame = view.bounds
+      spinner.center = CGPoint(x: view.bounds.midX, y: view.bounds.midY)
+    }
+
     override func viewDidAppear(_ animated: Bool) {
       super.viewDidAppear(animated)
-      webView?.webView.load(URLRequest(url: url))
+      webView.load(URLRequest(url: url))
     }
 
     // MARK: - Actions
@@ -106,13 +128,13 @@
 
     func webView(_ webView: WKWebView,
                  didStartProvisionalNavigation navigation: WKNavigation!) {
-      self.webView?.spinner.isHidden = false
-      self.webView?.spinner.startAnimating()
+      spinner.isHidden = false
+      spinner.startAnimating()
     }
 
     func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-      self.webView?.spinner.isHidden = true
-      self.webView?.spinner.stopAnimating()
+      spinner.isHidden = true
+      spinner.stopAnimating()
     }
 
     func webView(_ webView: WKWebView, didFail navigation: WKNavigation!,
